refactor: extract date input component
