+++ projects/client/src/lib/components/date-time/_internal/DateInputProps.ts
@@ -1,0 +1,7 @@
+import type { DatePickerProps } from './DatePickerProps.ts';
+
+export type DateInputType = 'date' | 'datetime-local';
+
+export type DateInputProps = {
+  type: DateInputType;
+} & DatePickerProps;
