+++ projects/client/src/lib/components/date-time/_internal/DatePickerProps.ts
@@ -1,0 +1,7 @@
+export type DatePickerProps = {
+  value?: Date;
+  maxDate?: Date;
+  label: string;
+  onChange: (date?: Date) => void;
+  disabled?: boolean;
+};
