+++ projects/client/src/lib/components/date-time/_internal/formatDateInputValue.ts
@@ -1,0 +1,8 @@
+import { formatLocalDateTime } from '../../../utils/date/formatLocalDateTime.ts';
+
+export function formatDateInputValue(
+  date: Date | undefined,
+) {
+  if (!date) return undefined;
+  return formatLocalDateTime(date);
+}
