+++ projects/client/src/lib/utils/date/formatLocalDateTime.ts
@@ -1,0 +1,5 @@
+import { format } from 'date-fns/format';
+
+export function formatLocalDateTime(date: Date): string {
+  return format(date, "yyyy-MM-dd'T'HH:mm");
+}
